+++ tests/logical/test_schema2.py
@@ -6,7 +6,7 @@ import pytest
 
 from daft.datatype import DataType
 from daft.expressions2 import ExpressionsProjection, col
-from daft.logical.schema2 import Schema
+from daft.logical.schema import Schema
 from daft.table import Table
 
 DATA = {
